+++ agli_shield/HydroCtl.h
@@ -30,7 +30,7 @@ class HydroCtlClass{
     
     static void     update(void);
     
-    uint32_t getEcValue(uint8_t[]);
+    unsigned long getEcValue(uint8_t[]);
     uint32_t    getWaterTemp(uint8_t[]);
     void getSensorValues(uint8_t[]);
     int waterEmpty();
